Move function to __init__.py
